% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqsR.R
\docType{package}
\name{cqsR-package}
\alias{cqsR}
\alias{cqsR-package}
\title{cqsR: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
