% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsEasierModel.R
\name{easierSummaryByValue}
\alias{easierSummaryByValue}
\title{summary/predict model result based on variable's value}
\usage{
easierSummaryByValue(modelResult, varOneToExtract, varOneToExtractValues)
}
\description{
summary/predict model result based on variable's value
}
